api_segformer/requirements.txt
# requirements.txt - API SegFormer P9
Flask==2.3.2
flask-cors==4.0.0
Pillow==10.0.0
numpy==1.24.3
onnxruntime==1.22.0
transformers==4.30.2
torch==2.0.1
torchvision==0.15.2
gunicorn==21.2.0
requests==2.31.0
pytest==7.4.0